// src/app/@shared/interfaces/request-access.ts
export interface RequestAccess {
    v_dar_product_name: string;
    radar_productname: string;
    v_dar_dm_version: string;
    v_dar_createdby_fullname: string;
    v_dar_objectid_createdby: string;
    requesting_id: string;
    v_dar_objectid_reqfor: string;
    v_dar_copy_access_profile_of: string;
    companyName: string;
    v_dar_organization: string;
    v_dar_accessType: string;
    v_dar_ad_group: string;
    accessDateContS: string;
    chk_dar_terms_message: string;
    edm_access: string;
    edm_PI_personal_details: string;
    expected_other: string;
    if_other_please_specify: string;
    v_custom_control_group: string;
    v_custom_routing_rule_identifier: string;
    m_dar_businessJustification_message: string;
    m_dar_terms_message: string;
    m_show_pii_message: string;
    v_dar_accessReqClassification: string;
    v_dar_accessStage: string;
    v_dar_userType: string;
    v_dar_additionalInfoContS: string;
    v_dar_platform:string;
    v_dar_analytics_platform: string;
    v_dat_api_name:string;
    v_dar_post_database_schema:string;
    v_dar_post_table_names:string;
    v_dar_tbdp_is_this_request_fort_irm_pricing:string;
    v_dar_attributeDetails: string;
    v_dar_businessJustification: string;
    v_dar_databaseName: string;
    v_dar_databaseType: string;
    v_dar_dataDomainSelect: string;
    v_dar_environment: string;
    v_dar_expectedAvailability: string;
    v_dar_frequencyOfUse: string;
    v_dar_maxUsers: string;
    v_dar_OtherType: string;
    v_dar_questionsContS: string;
    v_dar_reportNames: string;
    v_dar_requestedDataContS: string;
    v_dar_sourceApplications: string;
    v_dar_tableNames: string;
    v_dar_workdayID: string;
    v_dat_data_access_type: string;
    v_edm_3rd_party_vendor: string;
    v_edm_option: string;
    v_edm_type_of_work: string;
    v_emd_PI_access: string;
    v_emd_specify_PII_needs: string;
    v_dar_DataReqDescription: string;
}